fix(overlay): guard ctx/cancel against Run/Stop race; fix test lint
The new startup coverage tests surfaced a real data race: Overlay.Run
is launched in a goroutine and lazily initialises o.cancel, while a
concurrent Stop (error-path teardown) reads it. Guard both with a
dedicated lifecycleMu.

Also fix golangci-lint findings in the coverage tests: an ineffectual
buf assignment (ineffassign) and two unnecessary byte() conversions
(unconvert).

+++ shamap/compare_cov_test.go
@@ -1055,8 +1055,8 @@ func TestCmpCompareMaxCountOnInnerNodes(t *testing.T) {
 	var entries []cmp_entry
 	for i := byte(0); i < 30; i++ {
 		var k [32]byte
-		k[0] = byte(i * 7) // spread across different nibbles
-		k[1] = byte(i)
+		k[0] = i * 7 // spread across different nibbles
+		k[1] = i
 		entries = append(entries, cmp_entry{k, cmp_val(i)})
 	}
 	m1, _ := New(TypeState)
